requirements.txt
addict==2.4.0
appdirs==1.4.4
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
cityscapesScripts==2.2.0
coloredlogs==15.0.1
cycler==0.10.0
filelock==3.9.0
gdown==4.2.0
humanfriendly==9.2
idna==3.4
imageio==2.24.0
kiwisolver==1.2.0
kornia==0.5.8
matplotlib==3.4.2
mmcv-full==1.3.7
numpy==1.19.3
opencv-python==4.4.0.46
pandas==1.1.3
Pillow==9.5.0
prettytable==2.1.0
pyparsing==2.4.7
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2020.1
PyYAML==5.4.1
requests==2.28.2
scipy==1.6.3
seaborn==0.11.1
setuptools==49.2.1
six==1.16.0
soupsieve==2.3.2.post1
timm==0.3.2
torch==1.7.1+cu110
torchvision==0.8.2+cu110
tqdm==4.48.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.14
wcwidth== 0.2.5
yapf==0.31.0

